magistracy/storage-and-databases/load-tarantool-cluster/requirements.txt
tarantool
python-dotenv
